src/app/autogen-agents-customer-system/requirements.txt
# AutoGen多智能体客服系统依赖包

# AutoGen框架 - 核心智能体框架
pyautogen>=0.2.0

# FastAPI - Web API框架
fastapi>=0.104.0
uvicorn>=0.24.0

# 数据验证和序列化
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 异步支持
aiofiles>=23.2.0
httpx>=0.25.0

# 环境变量管理
python-dotenv>=1.0.0

# 时间处理
python-dateutil>=2.8.2

# OpenAI API
openai>=1.3.0

# CLI界面增强
rich>=13.7.0
colorama>=0.4.6

# 重试机制
tenacity>=8.2.0

# 测试框架
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 代码质量
black>=23.11.0